requirements.txt
torch
torchvision
timm
gradio
numpy
pillow